requirements.txt
certstore>=0.0.2
click>=8.0.1
keyring>=23.1.0
pyTenable>=1.3.3
oauthlib>=3.1.1
pandas>=1.3.2
tabulate>=0.8.9
